fix badge and Checkbox tests

+++ jdi-bdd/src/main/java/com/epam/jdi/bdd/stepdefs/RangeSteps.java
@@ -20,11 +20,11 @@ public class RangeSteps {
         range(name).setupValue(value);
     }
     @Then("^the \"([^\"]*)\" range value less or equal to (\\d+)$")
-    public void rangeLessOrEqualTo(String name, int value) {
+    public void rangeLessOrEqualTo(String name, double value) {
         range(name).is().value(lessThanOrEqualTo(value));
     }
     @Then("^the \"([^\"]*)\" range value greater or equal to (\\d+)$")
-    public void rangeGreaterOrEqualTo(String name, int value) {
+    public void rangeGreaterOrEqualTo(String name, double value) {
         range(name).is().value(greaterThanOrEqualTo(value));
     }
 }
